refactor(*): add base select query
Ref: #1194
