+++ applications/packages/core/core.orm/package.json
@@ -10,7 +10,7 @@
             "types": "./dist/index.d.ts"
         }
     },
-    "version": "4.3.0",
+    "version": "4.3.1",
     "license": "LGPL-3.0-only",
     "scripts": {
         "lint": "eslint .",
@@ -24,7 +24,8 @@
         "@wraithlight/core.nullable": "*",
         "@wraithlight/framework.primitive": "*",
         "@wraithlight/framework.type-utils": "*",
-        "mysql2": "3.2.0"
+        "mysql2": "3.2.0",
+        "@wraithlight/framework.nullable": "*"
     },
     "devDependencies": {
         "@regru/eslint-plugin-prefer-early-return": "1.0.0",
