+++ applications/packages/core/core.orm/src/context/query/_internal/query-context.model.ts
@@ -1,4 +1,4 @@
 export interface QueryConcatResult {
-    sql: string;
-    values: Array<string>;
+  sql: string;
+  values: Array<string>;
 }
